"gray bug"

+++ src/rgb_to_gray.cpp
@@ -9,9 +9,9 @@ void rgb_to_gray(
   gray.resize(height*width);
   ////////////////////////////////////////////////////////////////////////////
   // Add your code here
-  for (int i = 0; i < gray.size(); i++) {
+  for (int i = 0; i < (int)gray.size(); i++) {
 	  // more precise formula from stackoverflow
-	  gray[i] = (int)((double)rgb[i * 3] * 0.2126 + (double)rgb[i * 3 + 1] * 0.7152 + (double)rgb[i * 3 + 2] * 0.0722);
+	  gray[i] = (int)( (double)rgb[i * 3] * 0.2126 + (double)rgb[i * 3 + 1] * 0.7152 + (double)rgb[i * 3 + 2] * 0.0722 );
   }
   ////////////////////////////////////////////////////////////////////////////
 }
